separated editor code, and added preprocessor directives
