+++ Assets/Blocks/Scripts/QueueReadingMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cca2da0a42111954195d81d9da8f6a08
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
