+++ Assets/Map/Turtle/MovementMenu.cs
@@ -1,3 +1,4 @@
+#if UNITY_EDITOR
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
@@ -58,3 +59,4 @@ public class MovementMenu : Editor
         }
     }
 }
+#endif
