+++ Assets/Scripts/DevTools/ThumbnailScreenshots.cs
@@ -1,3 +1,4 @@
+#if UNITY_EDITOR
 using UnityEngine;
 using UnityEditor;
 using System;
@@ -41,3 +42,4 @@ public class ThumbnailScreenshots : MonoBehaviour
         
     }
 }
+#endif
